chore(zipLatestFrom): rename with -> pick
