+++ packages/core/src/index.js
@@ -11,9 +11,9 @@ export { periodic } from './source/periodic'
 
 export { newStream } from './source/newStream'
 
-import { zipArrayValues as _zipArrayValues, withArrayValues as _withArrayValues } from './combinator/withArrayValues'
+import { zipArrayValues as _zipArrayValues, pickArrayValues as _pickArrayValues } from './combinator/zipArrayValues'
 export const zipArrayValues = curry3(_zipArrayValues)
-export const withArrayValues = curry2(_withArrayValues)
+export const pickArrayValues = curry2(_pickArrayValues)
 
 // -----------------------------------------------------------------------
 // Observing
@@ -100,9 +100,9 @@ export const combineArray = curry2(_combineArray)
 // -----------------------------------------------------------------------
 // Sampling
 
-import { withLatestFrom as _withLatestFrom, zipLatestFrom as _zipLatestFrom } from './combinator/withLatestFrom'
+import { pickLatestFrom as _pickLatestFrom, zipLatestFrom as _zipLatestFrom } from './combinator/zipLatestFrom'
 
-export const withLatestFrom = curry2(_withLatestFrom)
+export const pickLatestFrom = curry2(_pickLatestFrom)
 export const zipLatestFrom = curry3(_zipLatestFrom)
 
 // -----------------------------------------------------------------------
