+++ packages/core/src/index.js.flow
@@ -77,8 +77,8 @@ declare export function merge <A> (s1: Stream<A>, s2: Stream<A>): Stream<A>
 declare export function merge <A> (s1: Stream<A>): (s2: Stream<A>) => Stream<A>
 declare export function mergeArray <A> (ss: Array<Stream<A>>): Stream<A>
 
-declare export function withLatestFrom <A, B> (values: Stream<A>, stream: Stream<B>): Stream<A>
-declare export function withLatestFrom <A, B> (values: Stream<A>): (stream: Stream<B>) => Stream<A>
+declare export function pickLatestFrom <A> (values: Stream<A>, stream: Stream<any>): Stream<A>
+declare export function pickLatestFrom <A> (values: Stream<A>): (stream: Stream<any>) => Stream<A>
 
 declare export function zipLatestFrom <A, B, C> (f: (A, B) => C, values: Stream<A>, stream: Stream<B>): Stream<C>
 declare export function zipLatestFrom <A, B, C> (f: (A, B) => C): (values: Stream<A>, stream: Stream<B>) => Stream<C>
