+++ packages/core/test/combinator/zipArrayValues-test.js
@@ -2,12 +2,12 @@
 import { describe, it } from 'mocha'
 import { eq, fail } from '@briancavalier/assert'
 
-import { zipArrayValues, withArrayValues } from '../../src/combinator/withArrayValues'
+import { zipArrayValues, pickArrayValues } from '../../src/combinator/zipArrayValues'
 import { empty } from '../../src/source/empty'
 
 import { collectEventsFor, makeEvents } from '../helper/testEnv'
 
-describe('withArrayValues', () => {
+describe('pickArrayValues', () => {
   describe('zipArrayValues', () => {
     it('should be empty for empty array', () => {
       const s = zipArrayValues(fail, [], makeEvents(1, 1))
@@ -45,21 +45,21 @@ describe('withArrayValues', () => {
     })
   })
 
-  describe('withArrayValues', () => {
+  describe('pickArrayValues', () => {
     it('should be empty for empty array', () => {
-      const s = withArrayValues([], makeEvents(1, 1))
+      const s = pickArrayValues([], makeEvents(1, 1))
       return collectEventsFor(1, s).then(eq([]))
     })
 
     it('should be empty for empty stream', () => {
-      const s = withArrayValues([1, 2, 3], empty())
+      const s = pickArrayValues([1, 2, 3], empty())
       return collectEventsFor(1, s).then(eq([]))
     })
 
     it('should contain array values when more events than values', () => {
       const a = ['a', 'b', 'c']
       const n = a.length + 1
-      const s = withArrayValues(a, makeEvents(1, n))
+      const s = pickArrayValues(a, makeEvents(1, n))
 
       return collectEventsFor(n, s)
         .then(eq([
@@ -72,7 +72,7 @@ describe('withArrayValues', () => {
     it('should contain array values when more values than events', () => {
       const a = ['a', 'b', 'c']
       const n = a.length - 1
-      const s = withArrayValues(a, makeEvents(1, n))
+      const s = pickArrayValues(a, makeEvents(1, n))
 
       return collectEventsFor(a.length, s)
         .then(eq([
