+++ packages/core/type-definitions/combinator/zipArrayValues.d.ts
